+++ CotEditor/Sources/SyntaxStyle.swift
@@ -433,7 +433,7 @@ extension SyntaxStyle {
             // expand highlight area if the character just before/after the highlighting area is the same color
             if let layoutManager = textStorage.layoutManagers.first {
                 var start = highlightRange.location
-                var end = highlightRange.max - 1
+                var end = highlightRange.max
                 var effectiveRange = NSRange.notFound
                 
                 if start <= bufferLength {
@@ -450,7 +450,7 @@ extension SyntaxStyle {
                                                     atCharacterIndex: end,
                                                     longestEffectiveRange: &effectiveRange,
                                                     in: wholeRange) != nil {
-                    end = effectiveRange.max - 1
+                    end = effectiveRange.max
                 }
                 
                 highlightRange = NSRange(location: start, length: end - start)