+++ CHANGELOG.md
@@ -5,6 +5,11 @@ Change Log
 develop
 --------------------------
 
+### Improvements
+
+- Exclude file extension from the initial selection in document save panel.
+
+
 ### Fixes
 
 - [beta] Fix syntax highlight parsing range while editing.